+++ zone/spawn2.cpp
@@ -445,42 +445,6 @@ void Spawn2::DeathReset(bool realdeath)
 	}
 }
 
-inline std::vector<uint32_t> GetLootdropIds(const std::vector<ZoneStateSpawnsRepository::ZoneStateSpawns> &spawn_states) {
-	LogInfo("Loading lootdrop ids for zone state spawns");
-
-	std::vector<uint32_t> lootdrop_ids;
-	for (auto &s: spawn_states) {
-		if (s.loot_data.empty()) {
-			continue;
-		}
-
-		LootStateData l{};
-		try {
-			std::stringstream ss;
-			{
-				ss << s.loot_data;
-				cereal::JSONInputArchive ar(ss);
-				l.serialize(ar);
-			}
-		}
-		catch (const std::exception &e) {
-			LogError("Failed to load loot state data for spawn2 [{}]: {}", s.id, e.what());
-			continue;
-		}
-
-		for (auto &e: l.entries) {
-			// make sure it isn't already in the list
-			if (std::find(lootdrop_ids.begin(), lootdrop_ids.end(), e.lootdrop_id) == lootdrop_ids.end()) {
-				lootdrop_ids.push_back(e.lootdrop_id);
-			}
-		}
-	}
-
-	LogInfo("Loaded [{}] lootdrop id(s)", lootdrop_ids.size());
-
-	return lootdrop_ids;
-}
-
 bool ZoneDatabase::PopulateZoneSpawnList(uint32 zoneid, LinkedList<Spawn2*> &spawn2_list, int16 version) {
 
 	std::unordered_map<uint32, uint32> spawn_times;
@@ -561,107 +525,7 @@ bool ZoneDatabase::PopulateZoneSpawnList(uint32 zoneid, LinkedList<Spawn2*> &spa
 		)
 	);
 	if (!spawn_states.empty()) {
-		LogInfo("Loading zone state spawns for zone [{}] spawns [{}]", zone_name, spawn_states.size());
-
-		std::vector<uint32_t> lootdrop_ids = GetLootdropIds(spawn_states);
-		zone->LoadLootDrops(lootdrop_ids);
-
-		for (auto &s: spawn_states) {
-			if (s.spawngroup_id == 0) {
-				continue;
-			}
-
-			if (s.is_corpse) {
-				continue;
-			}
-
-			uint32 spawn_time_left = 0;
-			if (spawn_times.count(s.spawn2_id) != 0) {
-				spawn_time_left = spawn_times[s.spawn2_id];
-				LogInfo("Spawn2 [{}] Respawn time left [{}]", s.spawn2_id, spawn_time_left);
-			}
-
-			// load from spawn2_disabled
-			bool spawn_enabled = true;
-
-			// check if spawn is disabled
-			for (auto &ds: disabled_spawns) {
-				if (ds.spawn2_id == s.spawn2_id) {
-					spawn_enabled = !ds.disabled;
-				}
-			}
-
-			auto new_spawn = new Spawn2(
-				s.spawn2_id,
-				s.spawngroup_id,
-				s.x,
-				s.y,
-				s.z,
-				s.heading,
-				s.respawn_time,
-				s.variance,
-				spawn_time_left,
-				s.grid,
-				(bool) s.path_when_zone_idle,
-				s.condition_id,
-				s.condition_min_value,
-				spawn_enabled,
-				(EmuAppearance) s.anim
-			);
-
-			if (spawn_time_left == 0) {
-				new_spawn->SetCurrentNPCID(s.npc_id);
-				new_spawn->SetLootStateData(s.loot_data);
-			}
-
-			spawn2_list.Insert(new_spawn);
-			new_spawn->Process();
-		}
-
-		// dynamic spawns, quest spawns, triggers etc.
-		for (auto &s: spawn_states) {
-			if (s.spawngroup_id > 0) {
-				continue;
-			}
-
-			auto npc_type = content_db.LoadNPCTypesData(s.npc_id);
-			if (npc_type == nullptr) {
-				LogError("Failed to load NPC type data for npc_id [{}]", s.npc_id);
-				continue;
-			}
-
-			auto npc = new NPC(
-				npc_type,
-				nullptr,
-				glm::vec4(s.x, s.y, s.z, s.heading),
-				GravityBehavior::Water
-			);
-
-			npc->AddLootStateData(s.loot_data);
-
-//			npc->AddLootTable();
-//			if (npc->DropsGlobalLoot()) {
-//				npc->CheckGlobalLootTables();
-//			}
-
-			if (s.grid) {
-				npc->AssignWaypoints(s.grid);
-			}
-
-			entity_list.AddNPC(npc, true, true);
-
-			if (s.is_corpse) {
-				auto decay_time = s.decay_in_seconds * 1000;
-				if (decay_time > 0) {
-					npc->SetQueuedToCorpse();
-					npc->SetCorpseDecayTime(decay_time);
-				} else {
-					npc->Depop();
-					continue;
-				}
-			}
-		}
-
+		zone->LoadZoneState(spawn_states, spawn_times, disabled_spawns);
 	} else {
 		for (auto &s: spawns) {
 			uint32 spawn_time_left = 0;
